MainVisual as an independent component
